Update README.md
This is a machine learning project that incorporates a graphical user interface (GUI) created using the Streamlit framework. The purpose of this project is to detect the quality of rice grains by utilizing their physical attributes, such as length, width, color, texture, and weight. The application also showcases sample images of the rice grains after the prediction process